broke logic into separate files
